64-bit transition part 12: ksh93 lexing, parsing and subshells
This is the twelfth of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The lexing and parsing components in lex.c, parse.c, fcin.c
  and trestore.c.
  - Added an assert() after stkalloc() to fix the following
    LTO warning:
    warning: 'memset' writing 60 bytes into a region of size 0 overflows the destination [-Wstringop-overflow=]
      1819 |                         memset(ioq,0,sizeof(*ioq));
           |                         ^
    lto1: note: destination object is likely at address zero
- Minor fixes for the virtual subshell mechanism.
- The code underlying shcomp(1), aka sh_tdump().
- A minor fix to a cast in sh_timeradd().

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
481 => 352 => 37 (progression from part 11 => part 12 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/cmd/ksh93/include/fcin.h
@@ -12,6 +12,7 @@
 *                                                                      *
 *                  David Korn <dgk@research.att.com>                   *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 #ifndef _FCIN_H
@@ -34,9 +35,9 @@ typedef struct _fcin
 	unsigned char	*fcptr;		/* pointer to next input char */
 	unsigned char	fcchar;		/* saved character */
 	short		fclen;		/* last multibyte char len */
-	void (*fcfun)(Sfio_t*,const char*,int,void*);	/* advance function */
+	void (*fcfun)(Sfio_t*,const char*,ptrdiff_t,void*);	/* advance function */
 	void		*context;	/* context pointer */
-	int		fcleft;		/* for multibyte boundary */
+	ptrdiff_t	fcleft;		/* for multibyte boundary */
 } Fcin_t;
 
 #if SHOPT_MULTIBYTE
@@ -46,9 +47,9 @@ typedef struct _fcin
 #   define fcmbget(x)	(fcget())
 #endif
 #define fcfile()	(_Fcin._fcfile)
-#define fcgetc()	(*_Fcin.fcptr++ ? (int)_Fcin.fcptr[-1] : fcfill())
-#define fcget()		((int)(*_Fcin.fcptr++))
-#define fcpeek(n)	((int)_Fcin.fcptr[n])
+#define fcgetc()	(*_Fcin.fcptr++ ? _Fcin.fcptr[-1] : fcfill())
+#define fcget()		(*_Fcin.fcptr++)
+#define fcpeek(n)	(_Fcin.fcptr[n])
 #define fcseek(n)	((char*)(_Fcin.fcptr+=(n)))
 #define fcfirst()	((char*)_Fcin.fcbuff)
 #define fclast()	((char*)_Fcin.fclast)
@@ -57,9 +58,9 @@ typedef struct _fcin
 #define fcsave(x)	(*(x) = _Fcin)
 #define fcrestore(x)	(_Fcin = *(x))
 extern int		fcfill(void);
-extern int		fcfopen(Sfio_t*);
+extern ssize_t		fcfopen(Sfio_t*);
 extern int		fcclose(void);
-void			fcnotify(void(*)(Sfio_t*,const char*,int,void*),void*);
+void			fcnotify(void(*)(Sfio_t*,const char*,ptrdiff_t,void*),void*);
 
 extern Fcin_t		_Fcin;		/* used by macros */
